requirements.txt
bidict==0.22.0
certifi==2021.10.8
charset-normalizer==2.0.12
dnspython==1.16.0
eventlet==0.30.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
python-dotenv==0.19.2
python-engineio==4.3.2
python-socketio==5.6.0
requests==2.27.1
six==1.16.0
urllib3==1.26.9
websocket-client==1.3.2
